Update docs to conform with command
The secrets command takes the -w (overWrite) flag when calling with '-s'.  This allows you to update a stored secret.  The existing doc has two examples of this, one using the correct flag 'w' and one using a flag that doesn't seem to exist 'o'.  This simple diff just updates the doc to match the code.